requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-telegram-bot==20.7
python-dotenv==1.0.0
anthropic==0.7.8
plotly==5.17.0
pandas==2.1.4
alembic==1.13.1